+++ src/ui/modals/mod.rs
@@ -13,6 +13,7 @@ use crate::{
 pub mod add_to_playlist;
 pub mod confirm_modal;
 pub mod confirm_queue_clear;
+pub mod decoders;
 pub mod keybinds;
 pub mod outputs;
 pub mod rename_playlist;